Создан шаблон обработки нажатия на кнопки в главном админском меню, а также фабрика колбеков и админская клавиатура
