+++ bot/cbdata.py
@@ -1,0 +1,5 @@
+from aiogram.filters.callback_data import CallbackData
+
+class AdminMenuCallbackFactory(CallbackData, prefix='admin'):
+    action: str
+
