+++ bot/main.py
@@ -5,7 +5,7 @@ from aiogram.fsm.storage.memory import MemoryStorage
 from bot.config_reader import load_config
 
 from bot.handlers.default_commands import default_cmd_router
-from bot.handlers.admin import admin_router
+from bot.handlers.admin_handlers import admin_router
 
 
 async def main():
